Refactored functions to be modules increasing organization and testability.
