+++ README.md
@@ -1,38 +1,12 @@
-# create-svelte
-
-Everything you need to build a Svelte project, powered by [`create-svelte`](https://github.com/sveltejs/kit/tree/master/packages/create-svelte).
-
-## Creating a project
-
-If you're seeing this, you've probably already done this step. Congrats!
-
-```bash
-# create a new project in the current directory
-npm create svelte@latest
-
-# create a new project in my-app
-npm create svelte@latest my-app
-```
-
-## Developing
-
-Once you've created a project and installed dependencies with `npm install` (or `pnpm install` or `yarn`), start a development server:
-
-```bash
-npm run dev
-
-# or start the server and open the app in a new browser tab
-npm run dev -- --open
-```
-
-## Building
-
-To create a production version of your app:
-
-```bash
-npm run build
-```
-
-You can preview the production build with `npm run preview`.
-
-> To deploy your app, you may need to install an [adapter](https://kit.svelte.dev/docs/adapters) for your target environment.
+### What is this repository?
+This is currently the rebuild of my startpage project. My original project works but needs a good rebuild.  Startpage was originally created as a project to learn Svelte and JS. Things have changed and it's time to tidy it up. The original project does not have any testing whatsoever so this here we are trying to fix that.  
+### What are we going to test?
+#### Unit Tests
+- Develop unit tests for "+page.server.js" functions.
+- [x] Sanity CMS Client Tests for client configuration and imageRefToSanityCompatibleURL function.  (Rename is in progress for that)
+- [ ] Need Unit Tests for fetch weather from Weather API and the backup Open Weather Map
+- [ ] Need Tests for Air Quality
+- [ ] Need Tests for fetch from Sanity functions (bookmark groups, search providers, GIFS, and Background Image)
+#### End-End Tests
+- Use PlayWright to handle UI tests, more to come later.
+- Currently need to develop a Mock API for testing the backup API Weather API functions.
